Move RCN from lib file its own module
